pyproject.toml
[tool.poetry]
name = "mlflowartifactoryplugin"
version = "0.0.1"
description = ""
authors = ["Lucas A. Bonini <lucasamorimbonini@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
dohq-artifactory = "^0.8.0"
mlflow = "^1.23.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipykernel = "^6.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
